Move feature gates tests to correct classes
